+++ cmake/SearchForStuff.cmake
@@ -671,6 +671,17 @@ if (NOT WIN32)
 endif()
 
 ########################################
+# Find the Ignition_Transport library
+find_package(ignition-transport0 QUIET REQUIRED)
+if (NOT ignition-transport0_FOUND)
+  BUILD_WARNING ("Missing: Ignition Transport (libignition-transport0-dev)")
+else()
+  set (CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} ${IGNITION-TRANSPORT_CXX_FLAGS}")
+  include_directories(${IGNITION-TRANSPORT_INCLUDE_DIRS})
+  link_directories(${IGNITION-TRANSPORT_LIBRARY_DIRS})
+endif()
+
+########################################
 # Find QWT (QT graphing library)
 #find_path(QWT_INCLUDE_DIR NAMES qwt.h PATHS
 #  /usr/include